+++ Thousand.Core/Layout.cs
@@ -8,5 +8,5 @@ namespace Thousand.Layout
     public record LabelBlock(Font Font, Rect Bounds, string Content, IReadOnlyList<LabelLine> Lines);
     public record Shape(Rect Bounds, ShapeKind Kind, int CornerRadius, Stroke Stroke, Colour Fill); // XXX should not have Name
     public record Line(Point Start, Point End, Stroke Stroke);
-    public record Diagram(int Width, int Height, float Scale, Colour? Background, IReadOnlyList<Shape> Shapes, IReadOnlyList<LabelBlock> Labels, IReadOnlyList<Line> Lines);
+    public record Diagram(int Width, int Height, decimal Scale, Colour? Background, IReadOnlyList<Shape> Shapes, IReadOnlyList<LabelBlock> Labels, IReadOnlyList<Line> Lines);
 }
